requirements.txt
streamlit
pandas
plotly
pymongo
python-dotenv
spotipy
